Separate time and observation updates in demo app
Put the time update and drawing code on a different interval than
the measurement update. The filter seems to be handling this ok :).
When setting the measurement period too high (e.g. 10s), the filter
performance is horrible, which is expected as without regular
measurements it can only rely on its random walk process.
Setting the measurement rate to 1Hz produces satisfactory results.