Start with replay
Made a start with replay code. The current code is able to read
the contents of log.json (which is created using a to-be-published
python script based on recordings with minicom).
The code currently returns raw RSSI values instead of changes
in RSSI, so the filter output is garbage at the moment. There is
also a lot of code cleanup work remaining.