FIX numerical precision problems in alm observation step
Found and fixed some problems in the alm observation steps, that
were caused by approx_normpdf returning probabilities of NaN or
infinity. This does not seem to be a problem of the normpdf
itself, but of the variance estimation that sometimes returns
small negative variances (i.e. -1e-10). It is now fixed by
assuming a minimum variance in the ALM filter, but this will
likely be moved to device/filter.js later.